SarahPrime_WebApp/requirements.txt
flask
openai>=1.0.0
python-dotenv

